+++ src/main/java/com/refinedmods/refinedstorage/setup/ServerSetup.java
@@ -1,5 +1,7 @@
 package com.refinedmods.refinedstorage.setup;
 
+import com.refinedmods.refinedstorage.RS;
+import com.refinedmods.refinedstorage.command.CreateDiskCommand;
 import com.refinedmods.refinedstorage.command.PatternDumpCommand;
 import net.minecraft.command.Commands;
 import net.minecraftforge.event.RegisterCommandsEvent;
@@ -8,9 +10,7 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 public class ServerSetup {
     @SubscribeEvent
     public void onRegisterCommands(RegisterCommandsEvent e) {
-        e.getDispatcher().register(
-            Commands.literal("refinedstorage")
-                .then(PatternDumpCommand.register(e.getDispatcher()))
-        );
+        e.getDispatcher().register(Commands.literal(RS.ID).then(PatternDumpCommand.register()));
+        e.getDispatcher().register(Commands.literal(RS.ID).then(CreateDiskCommand.register()));
     }
 }